What's this? Online multiplayer?! With firebase?!
Add firebase code to index.html
Create multiplayer.js to hold the relevant code
Add dummy property to enpassant so it shows up in firebase because null
is treated as destroyed by firebase. Which makes sense because when
something doesn't exist, firebase returns null
Remove the commented out AJAX for my previous idea for multiplayer
Prevent undo and redo from being called while playing an online game
Freeze players when it isn't their turn to prevent double moves
Send the game to the DB when a move is made, derp
Grab the list of games waiting for a second player when the connect
button is pressed.
Synchronise that list with firebase using the correct event listeners
Add div to hide undo/redo and AI stuff when online
Add form for creating a new game. Very simple.
Add div for the list of games awaiting second players
Turn the nulls into negative ones and vice versa for storage in firebase
Load in new data when the data changes and redraw the board
Modify HTML to reflect which side one is when playing online
When the first player moves the game from the waiting list to the
playing list, the second player needs to grab the new game reference and
delete the old one.
When a second player joins, the first player needs to create a game in
the playing list and put its id in the waiting list's game
Creating a game requires a name and a clean board and is fairly simple
Joining a game is pretty simple. Some AJAX and changing the number of
players to two.
LOL Massive Commit